Segundo Commit

+++ src/main/java/persistencia/claseUsuarioJpaController.java
@@ -66,7 +66,7 @@ public class claseUsuarioJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
